Day06

+++ 2021/AdventOfCode.2021/AdventOfCode.2021.ConsoleApp/Days/Day01/Solver.cs
@@ -4,7 +4,7 @@ public class Solver : ISolver
 {
     private const string NEWLINE = "\n";
 
-    public int Solution01(string input)
+    public long Solution01(string input)
     {
         var values = input.Split(NEWLINE);
 
@@ -23,7 +23,7 @@ public class Solver : ISolver
         return increased;
     }
 
-    public int Solution02(string input)
+    public long Solution02(string input)
     {
         var values = input.Split(NEWLINE);
