+++ 2021/AdventOfCode.2021/AdventOfCode.2021.ConsoleApp/Days/Day02/Solver.cs
@@ -11,7 +11,7 @@ public class Solver : ISolver
     private const int ACTION = 0;
     private const int VALUE = 1;
 
-    public int Solution01(string input)
+    public long Solution01(string input)
     {
         var lines = input.Split(NEWLINE);
 
@@ -41,7 +41,7 @@ public class Solver : ISolver
         return forward * dept;
     }
 
-    public int Solution02(string input)
+    public long Solution02(string input)
     {
         var lines = input.Split(NEWLINE);
